+++ openrem/remapp/tests/test_dicom_qr.py
@@ -367,6 +367,38 @@ class QRPhilipsCT(TestCase):
         self.assertEqual(rst1.dicomqrrspseries_set.all().count(), 1)
         self.assertEqual(rst1.dicomqrrspseries_set.all()[0].series_description, u"radiation dose report")
 
+    def test_modalities_in_study_generation(self):
+        """
+        Testing that ModalitiesInStudy is generated if not returned by remote C-Find SCP
+        """
+        from collections import Counter
+        from remapp.netdicom.qrscu import _generate_modalities_in_study
+
+        query = DicomQuery.objects.get()
+        rst1 = query.dicomqrrspstudy_set.all()[0]
+
+        # Add in a fourth series with modality SR
+        rst1s4 = DicomQRRspSeries.objects.create(dicom_qr_rsp_study=rst1)
+        rst1s4.query_id = query.query_id
+        rst1s4.series_instance_uid = uuid.uuid4()
+        rst1s4.modality = u"SR"
+        rst1s4.series_number = 999
+        rst1s4.series_description = u"radiation dose report"
+        rst1s4.number_of_series_related_instances = 1
+        rst1s4.save()
+
+        # Delete the modalities in study data
+        rst1.set_modalities_in_study(None)
+        rst1.save()
+
+        _generate_modalities_in_study(rst1)
+
+        # Modalities in study should now be available again
+        self.assertEqual(Counter(rst1.get_modalities_in_study()), Counter([u'CT', u'SC', u'SR']))
+
+
+
+
 
 class ResponseFiltering(TestCase):
     """