+++ docs/changes.rst
@@ -4,6 +4,7 @@ OpenREM version history
 
 0.8.0b5 (2018-05-xx)
 --------------------
+* `#626`_  DICOM: isolated the generate modalities in study function and added testing
 * `#625`_  Imports: now using event level UIDs to process continued, cumulative and duplicate RDSRs
 * `#624`_  Charts: removed filter link on number of events histogram as it was not functioning correctly
 * `#623`_  Imports: changed name of Toshiba image based extractor routine